fixing progress indicator

+++ src/lib/paraFetch.c
@@ -187,7 +187,7 @@ ssize_t sinceLastStatus = 0;
 char *dateString = "";
 int star = 1;  
 int starMax = 20;  
-int starStep = 1;
+off_t starStep = 1;
 // TODO handle case-sensitivity of protocols input
 if (startsWith("http://",url) || startsWith("https://",url))
     {
